+++ internal/presentation/rest/handlers/get_orders/schema.go
@@ -17,10 +17,10 @@ func (h) Parse(ctx *gin.Context) (request, error) {
 }
 
 type orderItem struct {
-	Number     string `json:"number"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual,omitempty"`
-	UploadedAt string `json:"uploaded_at"`
+	Number     string  `json:"number"`
+	Status     string  `json:"status"`
+	Accrual    float64 `json:"accrual,omitempty"`
+	UploadedAt string  `json:"uploaded_at"`
 }
 
 type response = []orderItem
